Add files via upload
Project 1: Start Your Business Intelligence Analyst Training

Overview of the program and definition of professional objectives, exploration of the profession, its main aspects, and tasks. Understanding the soft skills and hard skills required for the role of a competent business intelligence analyst. Lastly, the ability to self-assess in terms of skills.

Assessed Skills:

Adopt the posture of a Business Intelligence Analyst
Self-assess using the Business Intelligence Analyst framework
Identify the soft skills specific to the Business Intelligence Analyst profession